refactor: observer 모듈 chat에서 분리
