+++ src/shared/lib/functions/observer.test.ts
