+++ src/shared/lib/functions/observer.ts
